+++ R/calc_epidemic_onset.R
@@ -7,8 +7,10 @@
 #' @param c_closure POSIXct formatted date to start the model running the model
 #'  This is usually at canopy closure (Wolf)
 #' @param weather data.table, formatted with \code{\link{format_weather}}
-#' @param cultivar_sus character, susceptibility of the cultivar in "R" resistant,
-#'  "S" susceptible, "MR" moderately resistant etc.
+#' @param cultivar_sus numeric, susceptibility of the cultivar on the 'bsa'
+#'  https://www.bundessortenamt.de variety list.
+#'  Susceptibility rating must range between 1 and 9.
+#'  1 = resistant, 9 = susceptible. Default is 5.
 #'
 #' @return If the input weather is conducive for epidemic, the
 #'  function returns a POSIX_ct date when epidemic commences. If no epidemic
@@ -37,7 +39,7 @@ calc_epidemic_onset <- function(start,
   if(missing(end)) end <- as.Date(last(weather$times))
   if(missing(c_closure)){
     warning("'c_closure' not supplied, setting 'start' as canopy closure date")
-    start <- c_closure
+    c_closure <- start
   }
 
   c_closure <- as.POSIXct(c_closure,tz = "UTC")
@@ -46,15 +48,20 @@ calc_epidemic_onset <- function(start,
     if(is.na(cc)) return(NA)
     if(cc >= end) stop("'c_closure' is after last weather date")
 
-    w <- copy(weather[times > as.POSIXct(start) &
-                        times < (as.POSIXct(end) + 3600),][times >= as.POSIXct(cc)])
+    w <- copy(weather[times >= as.POSIXct(cc) &
+                        times < (as.POSIXct(end) + 3600),])
 
     daily_inf_val <- calc_DIV(dat = w)
 
-    div_cs <- daily_inf_val[first(which(cumsum(DIV) >cultivar_sus)),
+    div_cs <- daily_inf_val[first(which(cumsum(DIV) > calc_susceptibility(cultivar_sus))),
                             as.POSIXct(paste(Year,Month,Day,sep = "-"), tz = "UTC")]
 
-    if(length(div_cs) == 0) div_cs <- sum(daily_inf_val$DIV, na.rm = TRUE) / cultivar_sus
+
+    if(length(div_cs) == 0) {
+      warning("weather conditions not suitable for epidemic onset within timeframe
+      supplied.
+            NA returned.")
+      return(NA)}
     # calculate percentage
 
     return(div_cs[1])